% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countNAs.R
\name{countNAs}
\alias{countNAs}
\title{Format dates within data frame}
\usage{
countNAs(x)
}
\arguments{
\item{x}{Vector or list where to counts NAs}
}
\description{
Count NAs in a list or vector
}
\keyword{dateFormat}
