% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTooltip.R
\name{createTooltip}
\alias{createTooltip}
\title{Create tooltips}
\usage{
createTooltip(characterVector)
}
\arguments{
\item{characterVector}{Named character vector}
}
\description{
Create Tooltips in HTML format with appropriate labels and breaks
}
\keyword{HTML,}
