% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateFormate.R
\name{dateFormate}
\alias{dateFormate}
\title{Format dates within data frame}
\usage{
dateFormate(df, Column, originalFormat = "\%Y-\%m-\%d",
  newFormat = "\%m/\%d/\%Y")
}
\arguments{
\item{df}{A data frame}

\item{originalFormat}{Specify which format the orignal columns have (needs to be consistent)}

\item{newFormat}{Specify which format to convert to}

\item{column}{A column name or a vector of column names}
}
\description{
This function takes names of columns in a data frame and converts the date format
}
\keyword{dateFormat}
