% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteNaRows.R
\name{deleteNaRows}
\alias{deleteNaRows}
\title{Delelte rows with a lot of NAs}
\usage{
deleteNaRows(df, n = nrow(df))
}
\arguments{
\item{df}{A data frame}

\item{n}{Cuttoff value for number of values that should not be NA (excluding n)}
}
\description{
Keep only rows that have a certain number of non NA values
}
\keyword{NA}
