% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffDays.R
\name{diffDays}
\alias{diffDays}
\title{Difference in days}
\usage{
diffDays(x, y, format = "\%m/\%d/\%Y")
}
\arguments{
\item{x}{A date}

\item{format}{Character vector of date format used, defaults to "%m/%d/%Y"}

\item{n}{Another date}
}
\description{
Calculates the difference in daays of two dates
}
\keyword{difference}
\keyword{time,}
