% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicateCommaSeparatedRows.R
\name{duplicateCommaSeparatedRows}
\alias{duplicateCommaSeparatedRows}
\title{For cells to be duplicated}
\usage{
duplicateCommaSeparatedRows(df, commaSeparatedColumn)
}
\arguments{
\item{df}{The data frame containing the column with values to be separated}

\item{commaSeparatedColumn}{Column with the values}
}
\description{
Function that can duplicate Rows based on comma separated values in one column
}
