% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entryQueryDb.R
\name{entryQueryDb}
\alias{entryQueryDb}
\title{Query data base with large lists}
\usage{
entryQueryDb(sql_frame, queryEntry, data = NULL)
}
\arguments{
\item{sql_frame}{The character string that represents the frame for the R object to be queried}

\item{queryEntry}{Character or numeric vector that represents te list of search terms}

\item{ORALCE_UID}{Oracle data base user ID, set up as system variable}

\item{ORALCE_PASS}{Oracle data base password, set up as system variable}

\item{CON_STRING}{Oracle data base connections string, set up as system variable}
}
\description{
For queries with lists of more than 1000 length, this functions splits the list, does the query and rejoins to return the complete result
}
