% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelDate.R
\name{excelDate}
\alias{excelDate}
\title{Adds an apostrophe to a column so it does not get automatically formatted in excel}
\usage{
excelDate(df, column)
}
\arguments{
\item{df}{A data frame}

\item{column}{Name of the column which should be preserved in current format}
}
\description{
This function preserves the format of a column for use in excel
}
\keyword{Date}
\keyword{excel,}
