% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDbConnection.R
\name{getDbConnection}
\alias{getDbConnection}
\title{Establish connection to Oracle data base}
\usage{
getDbConnection(user_id = Sys.getenv("ORACLE_UID"),
  password = Sys.getenv("ORACLE_PASS"),
  conString = Sys.getenv("CON_STRING"))
}
\arguments{
\item{ORALCE_UID}{Oracle data base user ID, set up as system variable}

\item{ORALCE_PASS}{Oracle data base password, set up as system variable}

\item{CON_STRING}{Oracle data base connections string, set up as system variable}
}
\description{
Establish a connection to Oracle data base. Should not be called directly but is used by query functions
}
