% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenize.R
\name{homogenize}
\alias{homogenize}
\title{Homogenize the column of a data frame in terms of formatting}
\usage{
homogenize(df, Column)
}
\arguments{
\item{df}{A data frame}

\item{Column}{Column that should be harmonized}
}
\description{
Harmonizes an unconsistently formatted  data frame column
}
\keyword{formatting}
