% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outersect.R
\name{outersect}
\alias{outersect}
\title{Elements of two vectors that are not found in both}
\usage{
outersect(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{Another vector}
}
\description{
Complementary to the standard intersect() function, this function returns elements which are not common between two vectors
}
\keyword{difference}
