% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeRowAndColumn.R
\name{relativeRowAndColumn}
\alias{relativeRowAndColumn}
\title{Get relative row and column indices}
\usage{
relativeRowAndColumn(df, rowName = "ABSOLUTE_RANGE",
  colName = "ABSOLUTE_COLUMN")
}
\arguments{
\item{df}{A data frame}

\item{rowName}{absolute row of larger grid}

\item{colName}{absolute column  of larger grid}
}
\description{
This function calculates the relative row and column indices of a grid that is the subset of a larger grid
}
\keyword{relativeRowAndColumn}
