% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeFromEndString.R
\name{removeFromEndString}
\alias{removeFromEndString}
\title{Remove last n characters from string}
\usage{
removeFromEndString(string, n)
}
\arguments{
\item{string}{original string}

\item{n}{number of positions to be removed from end}
}
\description{
Removes the last n positions from a string
}
\keyword{formatting}
