% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSpaceString.R
\name{removeSpacesString}
\alias{removeSpacesString}
\title{Remove spaces from string}
\usage{
removeSpacesString(string)
}
\arguments{
\item{string}{original string}
}
\description{
Removes all spaces from a string
}
\keyword{formatting}
